pyproject.toml
[project]
name = "up-pashkevich-test-task"
version = "0.1.0"
description = ""
authors = [
    { name = "Khalil", email = "h.sultanoff@yandex.ru" }
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-unfold (>=0.57.0,<0.58.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.9"
pytest = "^8.3.5"

[tool.ruff]
line-length = 100
exclude = ["migrations", "__pycache__"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C4"]
ignore = ["B008"]